@@ -9,23 +9,25 @@
     "jquery": "2.1.2",
     "lodash": "3.4.0",
     "moment": "2.9.0",
-    "react": "0.12.2",
-    "react-router": "~0.12.4",
+    "react": "0.13.1",
+    "react-router": "~0.13.1",
     "reflux": "0.2.7",
-    "validator": "3.23.0"
+    "validator": "3.23.0",
+    "whatwg-fetch": "^0.7.0"
   },
   "devDependencies": {
-    "babelify": "5.0.4",
     "babel-eslint": "1.0.12",
+    "babelify": "5.0.4",
     "browserify": "~7.1.0",
+    "classnames": "^1.2.0",
     "coffeeify": "~1.0.0",
     "del": "~0.1.3",
     "event-stream": "~3.0.5",
     "gulp": "~3.8.9",
     "gulp-connect": "2.2.0",
+    "gulp-eslint": "0.7.0",
     "gulp-foreach": "*",
     "gulp-if": "1.2.5",
-    "gulp-eslint": "0.7.0",
     "gulp-karma": "0.0.1",
     "gulp-notify": "~1.5.0",
     "gulp-rename": "1.2.0",
@@ -43,7 +45,6 @@
     "karma-sinon": "1.0.4",
     "phantomjs": "1.9.12",
     "pretty-hrtime": "*",
-    "reactify": "~1.0.0",
     "run-sequence": "~0.3.5",
     "vinyl-source-stream": "~0.1.1",
     "watchify": "~2.3.0"
@@ -51,7 +52,6 @@
   "browserify": {
     "transform": [
       "coffeeify",
-      "reactify",
       "babelify"
     ]
   },
@@ -63,20 +63,40 @@
   },
   "eslintConfig": {
     "env": {
-      "es6":true,
-      "browser":true,
-      "node":true
+      "es6": true,
+      "browser": true,
+      "node": true
     },
     "ecmaFeatures": {
-      "modules":true,
-      "jsx":true
+      "modules": true,
+      "jsx": true
     },
     "rules": {
-      "new-cap" : [2, {"newIsCap": true, "capIsNew": false}],
-      "no-unused-vars" : [1, { "vars": "all" }],
-      "no-underscore-dangle":false,
-      "key-spacing": [1, { "align":"colon" }],
-      "quotes": [1, "double", "avoid-escape"]
+      "new-cap": [
+        2,
+        {
+          "newIsCap": true,
+          "capIsNew": false
+        }
+      ],
+      "no-unused-vars": [
+        1,
+        {
+          "vars": "all"
+        }
+      ],
+      "no-underscore-dangle": false,
+      "key-spacing": [
+        1,
+        {
+          "align": "colon"
+        }
+      ],
+      "quotes": [
+        1,
+        "double",
+        "avoid-escape"
+      ]
     }
   }
 }
